+++ odoo/local-src/bso_dashboards/models/bso_dashboard_graph.py
@@ -5,8 +5,9 @@
 import json
 from collections import OrderedDict, defaultdict
 
-from odoo import models, fields, api
 from odoo.tools.safe_eval import safe_eval
+
+from odoo import models, fields, api
 
 
 class OrderedDefaultDict(OrderedDict, defaultdict):
@@ -311,4 +312,12 @@ class BSODashboardGraph(models.Model):
             "view_mode": "form",
             "target": "new",
             "res_id": self.settings_id.id,
+        }
+
+    @api.multi
+    def action_print_xls_report(self):
+        self.ensure_one()
+        return {
+            'type': 'ir.actions.report.xml',
+            'report_name': 'bso.dashboard.report',
         }